refactor(desktop): pr-agent run 子系统归入 services/pr-agent + 命名对齐
把队列与执行两个类移入 services/pr-agent/ 子目录并统一命名（目录已表 pr-agent，去冗余前缀）：
- RunQueueService → RunQueue（services/pr-agent/run-queue.ts）
- PragentRunExecutor → RunExecutor（services/pr-agent/run-executor.ts）
- 新增 services/pr-agent/index.ts 桶导出：对外只暴露 RunQueue + 队列类型，RunExecutor 作子系统内部协作件不外暴露。
- ipc / context / agent-orchestrator 改从 ./services/pr-agent 引用 RunQueue。

纯改名 / 移动、行为不变；desktop typecheck/lint/build 全绿。
