+++ apps/desktop/src/main/services/pr-agent/index.ts
@@ -1,0 +1,6 @@
+/**
+ * pr-agent run 子系统：调度（RunQueue：并发 / 优先级 / 取消）+ 执行（RunExecutor，内部协作件，不外暴露）。
+ * 对外只暴露 RunQueue 与队列相关类型。
+ */
+export { RunQueue } from './run-queue.js';
+export type { QueueItem, RunPriority } from './run-queue.js';
